Time to refactor our code to use the slug in the URLs. I'll close up a few files and then open GenusController. The show page we just saw in our browser comes from showAction(). And yep, it has {genusName} in the URL. Gross.Change that to {slug}
And now, because slug is a property on the Genus entity, we don't need to manually query for it anymore. Instead, type-hint Genus as an argument.

Now, Symfony will do our job for us: I mean, query for the Genus automatically.

That means we can clean up a lot of this code. Just update the  variable below to dollargenus->getName()

We just Broke our App!

Cool! Except, we just broke our app! By changing the wildcard from {genusName} to {slug}, we broke any code that generates a URL to this route. How can we figure out where those spots are?