Let's talk about the famous, ManyToMany relationship. We already have a Genus entity and also a User entity. Before this tutorial, I updated the fixtures file. It still loads genuses, but it now loads two groups of users:
37 lines src/AppBundle/DataFixtures/ORM/fixtures.yml
... lines 1 - 22
AppBundle\Entity\User:
    user_{1..10}:
        email: weaverryan+<current()>@gmail.com
        plainPassword: iliketurtles
        roles: ['ROLE_ADMIN']
        avatarUri: <imageUrl(100, 100, 'abstract')>
    user.aquanaut_{1..10}:
        email: aquanaut<current()>@example.org
        plainPassword: aquanote
        isScientist: true
        firstName: <firstName()>
        lastName: <lastName()>
        universityName: <company()> University
        avatarUri: <imageUrl(100, 100, 'abstract')>

The first group consists of normal users, but the second group has an isScientist boolean field set to true. In other words, our site will have many users, and some of those users happen to be scientists.

That's not really important for the relationship we're about to setup, the point is just that many users are scientists. And on the site, we want to keep track of which genuses are being studied by which scientists, or really, users. So, each User may study many genuses. And each Genus, may be studied by many Users.

This is a ManyToMany relationship. In a database, to link the genus table and user table, we'll need to add a new, middle, or join table, with genus_id and user_id foreign keys. That isn't a Doctrine thing, that's just how it's done.

Mapping a ManyToMany in Doctrine

So how do we setup this relationship in Doctrine? It's really nice! First, choose either entity: Genus or User, I don't care. I'll tell you soon why you might choose one over the other, but for now, it doesn't matter. Let's open Genus. Then, add a new private property: let's call it dollar_genusScientists:

This could also be called users or anything else. The important thing is that it will hold the array of User objects that are linked to this Genus