Avoiding Duplicates
Let's do some experimenting! What if I duplicated the addGenusScientist() line?

Could this one new Genus be related to the same User two times? Let's find out!

Refresh the new page again. Alright! I love errors!

    Duplicate entry '12-11' for key 'PRIMARY'

So this is saying:

    Yo! You can't insert two rows into the genus_scientist table for the same genus and user.

And this is totally by design - it doesn't make sense to relate the same Genus and User multiple times. So that's great... but I would like to avoid this error in case this happens accidentally in the future.