Inside the loop, add an li with some styling. Then add a link. Why a link? Because before this course, I created a handy-dandy user show page. Copy the user_show route name, then use path(), paste the route, and pass it an id set to genusScientist.id, which we know is a User object. Then, genusScientist.fullName. Why fullName? If you look in the User class, I added a method called getFullName(), which puts the firstName and lastName together.
It's really not that fancy.

Time for a test drive! When we refresh, we get the header, but this Genus doesn't have any scientists. Go back to /genus/new to create a more interesting Genus. Click the link to view it. Boom! How many queries did we need to write to make this work? None! That's right - we are keeping lazy.

But now, click to go check out the user show page. What if we want to do the same thing here? How can we list all of the genuses that are studied by this User? Time to setup the inverse side of this relationship