Our goal is clear: list all of the genuses studied by this User. The Owning vs Inverse Side of a Relation
Back in our Doctrine Relations tutorial, we learned that every relationship has two different sides: a mapping, or owning side, and an inverse side. In that course, we added a GenusNote entity and gave it a ManyToOne relationship to Genus.
This is the owning side, and it's the only one that we actually needed to create.

If you look in Genus, we also mapped the other side of this relationship: a OneToMany back to GenusNote.

This is the inverse side of the relationship, and it's optional. When we mapped the inverse side, it caused no changes to our database structure. We added it purely for convenience, because we decided it sure would be fancy and nice if we could say ->getNotes() to automagically fetch all the GenusNotes for this Genus.

With a ManyToOne relationship, we don't choose which side is which: the ManyToOne side is always the required, owning side. And that makes sense, it's the table that holds the foreign key column, i.e. GenusNote has a genus_id column.
Owning and Inverse in ManyToMany

We can also look at our ManyToMany relationship in two different directions. If I have a Genus object, I can say:

    Hello fine sir: please give me all Users related to this Genus.

But if I have a User object, I should also be able to say the opposite:

    Good evening madame: I would like all Genuses related to this User.

The tricky thing about a ManyToMany relationship is that you get to choose which side is the owning side and which is the inverse side. And, I hate choices! The choice does have consequences.... but don't worry about that - we'll learn why soon.

Mapping the Inverse Side

Since we only have one side of the relationship mapped now, it's the owning side. To map the inverse side, open User and add a new property: . This will also be a ManyToMany with targetEntity set to Genus. But also add mappedBy='genusScientists'. That refers to the property inside of Genus.
Now, on that property, add inversedBy='studiedGenuses', which points back to the property we just added in User