Using the Inverse Side
And just by adding this new property, we are - as I love to say - dangerous.

Head into the user/show.html.twig template that renders the page we're looking at right now. Add a column on the right side of the page, a little 'Genuses Studied' header, then a ul. To loop over all of the genuses that this user is studying, just say for genusStudied in user.studiedGenuses. Don't forget the endfor.
Inside, add our favorite list-group-item and then a link. Link this back to the genus_show route, passing slug set to genusStudied.slug. Print out genusStudied.name.

But will it blend? I mean, will it work? Refresh!

Hot diggity dog! There are the three genuses that this User studies. We did nothing to deserve this nice treatment: Doctrine is doing all of the query work for us.

In fact, click the database icon on the web debug toolbar to see what the query looks like. When we access the property, Doctrine does a SELECT from genus with an INNER JOIN to genus_scientist where genus_scientist.user_id equals this User's id: 11. That's perfect! Thanks Obama