Back on the Genus page, I want to add a little 'X' icon next to each user. When we click that, it will make an AJAX call that will remove the scientist from this Genus.
How a ManyToMany Link is Removed

To link a Genus and a User, we just added the User object to the genusScientists property.
So guess what? To remove that link and delete the row in the join table, we do the exact opposite: remove the User from the genusScientists property and save. Doctrine will notice that the User is missing from that collection and take care of the rest.

Setting up the Template

Let's start inside the the genus/show.html.twig template. Add a new link for each user: give some style classes, and a special js-remove-scientist-user class that we'll use in JavaScript. Add a cute close icon.