The Remove Endpoint Setup
Inside, we need to make an AJAX call back to our app. Let's go set that up. Open GenusController and find some space for a new method: public function removeGenusScientistAction(). Give it an {dollarSign}Route() set to /genus/{genusId}/scientist/{userId}

You see, the only way for us to identify exactly what to remove is to pass both the genusId and the userId. Give the route a name like genus_scientist_remove. Then, add an {dollarSign}Method set to DELETE.

You don't have to do that last part, but it's a good practice for AJAX, or API endpoints. It's very clear that making this request will delete something. Also, in the future, we could add another end point that has the same URL, but uses the GET method. That would return data about this link, instead of deleting it.

Any who, add the genusId and userId arguments on the method.

Next, grab the entity manager with {dollarSign}this->getDoctrine()->getManager() so we can fetch both object.
Add {dollarSign}genus = {dollarSign}em->getRepository('AppBundle:Genus')->find({dollarSign}genusId)

I'll add some inline doc to tell my editor this will be a Genus object. And of course, if !{dollarSign}genus, we need to throw {dollarSign}this->createNotFoundException(): 'genus not found'.

Copy all of that boring goodness, paste it, and change the variable to {dollarSign}genusScientist. This will query from the User entity using . If we don't find a {dollarSign}genusScientist, say 'genus scientist not found'