Deleting the Link
Now all we need to do is remove the User from the Genus. We don't have a method to do that yet, so right below addGenusScientist(), make a new public function called removeGenusScientist() with a User argument.
Inside, it's so simple: {dollarSign}this->genusScientists->removeElement({dollarSign}user)
In other words, just remove the User from the array... by using a fancy convenience method on the collection. That doesn't touch the database yet: it just modifies the array.