Returning from the Endpoint
At the bottom, we still need to return a Response. But, there's not really any information we need to send back to our JavaScript... so I'm going to return a new Response with null as the content and a 204 status code.

This is a nice way to return a response that is successful, but has no content. The 204 status code literally means 'No Content'.

Now, let's finish this by hooking up the frontend.