Making the AJAX Call
Scroll back to the javascripts block. I'll paste a few lines of code here to get us started.

Ok, no big deal: the first line uses {dollarSign}(this), which is the link that was just clicked, and finds the .js-scientist-item li that is around it. We'll use that in a minute. The second chunk changes the fa-close icon into a loading spinner... ya know... because we deserve fancy things.