On the genus list page, I want to add a new column that prints the number of scientists each Genus has. That should be simple!
Open the genus/list.html.twig template. Add the new th for number of scientists.
Then down below, add the td, then say {{ genus.genusScientists|length }}

In other words:

Go out and get my array of genus scientists and count them!

The Lazy Collection Queries

But now click the Doctrine icon down in the web debug toolbar to see how the queries look on this page. This is really interesting: we have one query that's repeated many times: it selects all of the fields from user and then INNER JOINs over to genus_scientist WHERE genus_id equals 29, then, 25, 26 and 27.

When we query for the Genus, it does not automatically also go fetch all the related Users. Instead, at the moment that we access the genusScientists property, Doctrine queries all of the User data for that Genus. We're seeing that query for each row in the table.