That's everything! Head to the template: app/Resources/views/admin/genus/_form.html.twig. Head to the bottom and simply add the normal form_row(genusForm.genusScientists).
Guys, let's go check it out.
Choosing the Choice Label

Refresh! And... explosion!

    Catchable Fatal Error: Object of class User could not be converted to string.

Wah, wah. Our form is trying to build a checkbox for each User in the system... but it doesn't know what field in User it should use as the display value. So, it tries - and fails epicly - to cast the object to a string.