Back on the main part of the site, click one of the genuses, and then click one of the users that studies it. This page has a little edit button: click that. Welcome to a very simple User form.  Building the Field
Ok, same plan: add checkboxes so that I can choose which genuses are being studied by this User. Open the controller.
This uses UserEditForm, so go open that as well.

In buildForm(), we'll do the exact same thing we did on the genus form: add a new field called studiedGenuses - that's the property name on User that we want to modify

Keep going: use EntityType::class and then set the options: class set now to Genus::class to make Genus checkboxes. Then, multiple set to true, expanded set to true, and choice_label set to name to display that field from Genus