Open the template: user/edit.html.twig. At the bottom, use form_row(userForm.studiedGenuses).
That's it.

Try it! Refresh! Cool! This User is studying five genuses: good for them! Let's uncheck one genus, check a new one and hit Update.

t didn't Work!!! Inverse Relationships are Read-Only

This is the moment where someone who doesn't know what we're about to learn, starts to hate Doctrine relations.

Earlier, we talked about how every relationship has two sides. You can start with a Genus and talk about the genus scientist users related to it.

Or, you can start with a User and talk about its studied genuses.

Only one of these side - in this case the Genus - is the owning side. So far, that hasn't meant anything: we can easily read data from either direction. BUT! The owning side has one special power: it is the only side that you're allowed to change.

What I mean is, if you have a User object and you add or remove genuses from its studiedGenuses property and save... Doctrine will do nothing. Those changes are completely ignored.

And it's not a bug! Doctrine is built this way on purpose. The data about which Genuses are linked to which Users is stored in two places. So Doctrine needs to choose one of them as the official source when it saves. It uses the owning side.

For a ManyToMany relationship, we chose the owning side when we set the mappedBy and inversedBy options. The owning side is also the only side that's allowed to have the @ORM\JoinTable annotation.

This is a long way of saying that if we want to update this relationship, we must add and remove users from the  property on Genus.

Adding and removing genuses from the User object will do nothing. And that's exactly what our form just did.

No worries! We can fix this, with just a little bit of really smart code.