So, should we create a setStudiedGenuses() method like it suggested? Actually, no. Another option is to create adder & remover methods.
Create a public function addStudiedGenus() with a Genus argument.

In adder function, we'll do the same type of thing we did back in our Genus class.
After that, add the remover: public function removeStudiedGenus() also with a Genus argument.

Perfect!

Go back to the form. Uncheck one of the genuses and check a new one. When we submit, it should call addStudiedGenus() once for the new checkbox and removeStudiedGenus() once for the box we unchecked.

Ok, hit update! Hmm, it looked successful... but it still didn't actually work. And that's expected! We just setup a cool little system where the form component calls our adder and remover methods to update the studiedGenuses property. But... this hasn't really changed anything: we're still not setting the owning side of the relationship.

But, we're just one small step from doing that.