Ultimately, when we submit... we're still only updating the studiedGenuses property in User, which is the inverse side of this relationship.
So, nothing actually saves.

How can we set the owning side? Why not just do it inside the adder and remover methods? At the bottom of addStudiedGenus(), add {dollar_sign}genus->addGenusScientist({dollar_sign}this)

we just set the owning side of the relationship!

In removeStudiedGenus(), do the same thing: {dollar_sign}genus->removeGenusScientist({dollar_sign}this)

that's all we need to do. Go back to the form, uncheck a genus, check a genus and hit update. It's alive

We didn't need to add a lot of code to get this to work... but this situation has caused many developers to lose countless hours trying to get their relationship to save. To summarize: if you're modifying the inverse side of a relationship, set the by_reference form option to false, create an adder and remover function, and make sure you set the owning side of the relationship in each. That is it.