Synchronizing Both Sides
So, we're done! Well, technically we are done, but there is one last, tiny, teeny detail that I'd like to perfect. In Genus, when we call addGenusScientist(), it would be nice if we also updated this User to know that this Genus is now being studied by it. In other words, it would be nice if we called {dollar_sign}user->addStudiedGenus({dollar_sign}this).

I'm also going to add a note: this is not needed for persistence, but it might save you from an edge-case bug. Suppose we called {dollar_sign}genus->addGenusScientist() to link a User to the Genus. Then later, during the same request - that's important - we have that same User object, and we call getStudiedGenuses(). We would want the Genus that was just linked to be in that collection. This does that! We're guaranteeing that both sides of the relationship stay synchronized.

Do the same thing down in the remover: {dollar_sign}user->removeStudiedGenus({dollar_sign}this)