Updating the Existing Relationships
Now that the join table has an entity, we need to update the relationships in Genus and User to point to it. In Genus, find the genusScientists property. Guess what? This is not a ManyToMany to User anymore: it's now a OneToMany to GenusScientist. Yep, it's now the inverse side of the ManyToOne relationship we just added. That means we need to change inversedBy to mappedBy set to genus. And of course, targetEntity is GenusScientist.
You can still keep the fetch='EXTRA_LAZY': that works for any relationship that holds an array of items. But, we do need to remove the JoinTable: annotation: both JoinTable and JoinColumn can only live on the owning side of a relationship.

There are more methods in this class - like addGenusScientist() that are now totally broken. But we'll fix them later. In GenusScientist, add inversedBy set to the genusScientists property on Genus.

Finally, open User: we need to make the exact same changes here.

For studiedGenuses, the targetEntity is now GenusScientist, the relationship is OneToMany, and it's mappedBy the user property inside of GenusScientist

The OrderBy doesn't work anymore. Well, technically it does, but we can only order by a field on GenusScientist, not on User. Remove that for now.

Also add the inversedBy='studiedGenuses' to the user property in GenusScientist.