In some ways, not much just changed. Before, we had a genus_scientist table with genus_id and user_id columns. And... we still have that, just with two new columns.
But, in our app, a ton just changed. That's my nice way of saying: we just broke everything!
Collection of GenusScientists, not Users

For example, before, genusScientists was a collection of User objects, but now it's a collection of GenusScientist objects.

The same thing is true on User.
Wherever our code was using the studiedGenuses property - to get the collection or change it - well, that code is done broke.

Let's clean things up! And see some cool stuff along the way.

Creating new Join Entity Links

First, because we just emptied our database, we have no data. Open the fixtures file and temporarily comment-out the genusScientists property.
We can't simply set a User object on genusScientists anymore: this now accepts GenusScientist objects. We'll fix that in a second.

But, run the fixtures:

bin/console doctrine:fixtures:load