Let's start in the Genus class itself. Find getGenusScientists(). This method was lying! It does not return an array of User objects, it returns an array of GenusScientist objects!
In the template, when we loop over genus.genusScientists, genusScientist is not a User anymore. Update to genusScientist.user.fullName, and above, for the user_show route, change this to genusScientist.user.id.

Then, in the link, let's show off our new yearsStudied field: {{ genusScientist.yearsStudied }} then years.

We still need to fix the remove link, but let's see how it looks so far!

Refresh! It's way less broken! Well, until you click to view the user!