Updating the User Template
To fix this, start by opening User and finding getStudiedGenuses(). Change the PHPDoc to advertise that this now returns an array of GenusScientist objects.

Next, go fix the template: user/show.html.twig. Hmm, let's rename this variable to be a bit more clear: genusScientist, to match the type of object it is. Now, update slug to be genusScientist.genus.slug. And print genusScientist.genus.name.

Try it! Page is alive